requirements.txt
# Object Detection Model Requirements
# Core dependencies for training and inference

# Deep Learning Framework
torch>=1.9.0
torchvision>=0.10.0

# Computer Vision
timm>=0.6.0
opencv-python>=4.5.0

# Configuration Management
PyYAML>=6.0

# Data Processing
numpy>=1.21.0
Pillow>=8.3.0

# Visualization (optional)
matplotlib>=3.4.0
tensorboard>=2.7.0

# Utilities
tqdm>=4.62.0
